added: generic suites
