+++ tests-utils/src/main/scala/tech/beshu/ror/utils/containers/generic/ElasticsearchNodeWaitingStrategy.scala
@@ -14,7 +14,7 @@
  *    You should have received a copy of the GNU General Public License
  *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
  */
-package tech.beshu.ror.utils.containers
+package tech.beshu.ror.utils.containers.generic
 
 import java.util.concurrent.TimeUnit
 
@@ -87,12 +87,4 @@ class ElasticsearchNodeWaitingStrategy(esVersion: String,
       }
     }
   }
-}
-
-trait ElasticsearchNodeDataInitializer {
-  def initialize(esVersion: String, adminRestClient: RestClient): Unit
-}
-
-object NoOpElasticsearchNodeDataInitializer extends ElasticsearchNodeDataInitializer {
-  override def initialize(esVersion: String, adminRestClient: RestClient): Unit = {}
 }
