+++ tests-utils/src/main/scala/tech/beshu/ror/utils/containers/generic/EsWithoutRorPluginContainer.scala
@@ -14,7 +14,7 @@
  *    You should have received a copy of the GNU General Public License
  *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
  */
-package tech.beshu.ror.utils.containers
+package tech.beshu.ror.utils.containers.generic
 
 import cats.data.NonEmptyList
 import com.typesafe.scalalogging.StrictLogging
